fix: use alternative oxi binding to get filetype (#28)
* fix: use alternative oxi binding to get filetype

This solution works, but is quite hacky. It doesn't seem to break the
api for end users though, and currently works on nightly. It likely
works on 0.8 as well.

* refactor: use more idiomatic error handling instead of unwrap

Instead of causing an immediate panic when a filetype can't be found,
instead boil an error up for user presentation stating a filetype isn't
found. This ensures Neovim doesn't just up and die when a panic occurs.

* refactor: remove unused import